+++ src/constants.ts
@@ -6,5 +6,6 @@ export const CSV_MULTIVALUE_SEPARATOR = ';';
 
 
 export const RESOURCE_TYPES = {
-    connectorRule : "connector-rules"
+    connectorRule : "connector-rules",
+    identityAttribute : "identity-attributes",
 } as const