+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog is following the recommended format by [keepachangelog](https://k
 ### Added
 
 - Added transforms "RFC5646" and "Display Name" (cf. [#87](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/87))
+- Can create or delete an identity attribute (cf. [#83](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/83))
 
 ### Changed